Move `EntityVertexId` to `subgraph` module
